refactor: RequestLine 삭제 후 기능 HttpRequest 이동
- HttpRequest 클래스 자체에서 HTTP 요청 라인을 파싱 처리하도록 변경
- 매개변수를 맵으로 처리
